+++ package.json
@@ -7,7 +7,7 @@
   "repository": "https://github.com/probot/dco.git",
   "scripts": {
     "start": "probot run ./index.js",
-    "test": "mocha && standard"
+    "test": "jest && standard"
   },
   "engines": {
     "node": ">= 7.7.0",
@@ -15,14 +15,13 @@
   },
   "devDependencies": {
     "eslint-plugin-import": "^2.7.0",
-    "expect": "^1.20.2",
+    "jest": "^22.0.6",
     "localtunnel": "^1.8.3",
-    "mocha": "^3.2.0",
     "standard": "^10.0.3"
   },
   "standard": {
     "env": [
-      "mocha"
+      "jest"
     ]
   },
   "dependencies": {
